LayoutTests/http/tests/security/mixedContent/insecure-css-in-iframe.html

<html>
<body>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
}
</script>
<p>This test loads a secure iframe that loads an insecure style sheet.  We
should *not* trigger a mixed content callback because an active network
attacker cannot corrupt the HTTPS security origin.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css.html";
></iframe>
</body>
</html>
